pyproject.toml
[project]
name = "prisma-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.8",
    "jwt>=1.3.1",
    "prisma>=0.15.0",
    "python-dotenv>=1.0.1",
    "webauthn>=2.5.1",
]